modified Function and its subclasses

+++ src/Psf.cxx
@@ -53,7 +53,7 @@ double Psf::value(astro::SkyDir appDir, double energy,
 // Here we assume the m_scTimes are at regular intervals.
    double tstep = m_scData[1].time - m_scData[0].time;
    int indx;
-   indx = (time - m_scData[0].time)/tstep;
+   indx = static_cast<int>((time - m_scData[0].time)/tstep);
 
 // inclination wrt spacecraft z-axis in degrees
    double inc = srcDir.SkyDir::difference(m_scData[indx].zAxis)*180./M_PI;
